+++ CHANGELOG.md
@@ -12,10 +12,12 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 ### Added
 
 - We added an option to create entries directly from Bib(La)TeX sources to the 'Create New Entry' tool. [#8808](https://github.com/JabRef/jabref/issues/8808)
+- We added 'Make absolute' button which will convert the selected path to an absolute path in context of library properties. [#13031](https://github.com/JabRef/jabref/issues/13031)
 
 ### Changed
 
 - We merged the 'New Entry', 'Import by ID', and 'New Entry from Plain Text' tools into a single 'Create New Entry' tool. [#8808](https://github.com/JabRef/jabref/issues/8808)
+- We changed path output display to show the relative path with respect to library path in context of library properties. [#13031](https://github.com/JabRef/jabref/issues/13031)   
 
 ### Fixed